Connect elevator and rudder gauges
Rework aero for aileron - as the values are +/-10 degrees - however the roll rate is too slow so it would appear that these are combined values for the deflection (L+R) rather than actual position - so add a factor of two on the coeffs that use this.

Add priorty to fuel tanks - as the fwd/aft should be last.
Extra entry into EGT lookup for the way that the AB is modelled (>99.8% is augmented)