Ground spoilers and connect spoiler instrumentation
The logic for the ground spoilers was slightly wrong; partly because you can't have sub conditions in switches in JSBSim; so rework the ground spoilers deployed to use discrete elements.

Connect the spoilers to the appropriate gauge (in the fuel panel).

Move all of the spoiler modelling into the FDM (including -ve extension to meet flaps)