Use ALS shadows (via interior effect).
This isn't right as the shadows are scaled per model - so the scaling is the same on a gauge as it would be for the entire cockpit. Not sure how to fix this - it might not be possible and need a new scheme whereby the entire cockpit is within an overall group (merged) so that the shadows will work.

There would probably need to be two groups - one for the back lit elements and another for the rest. The important thing is that these elements are rendered via one effect shader so that the scaling works; or at least that's what I think at the moment.