Weapons handling;
Modifications to support the correct display on the MPCD for PACS.
AIM9 count.
AWG9 mods related to problems from submodel of AIM9 after loading.

This is nearly working; the AIM9 will be selected and the AIM9 object created; firing still not quite working.
Once this is working the next step will be the other missile types; probably using an adaptation of Fabien's M2000 missiles code