F-15: Added ground vehicles (tanker/cart).
Although not totally happy with having these attached directly to the model it is the best way I've found to make them follow the aircraft and be easily controllable. It might be that with a better understanding of submodels I could get this to work - but my attempts failed thus fair so this technique will suffice until something better comes along.

The larger version of the ground fuelling tanker isn't currently used; and possibly both aren't really that accurate but will do for now.