Added emergency gear.
When the handle is turned it will cut power to the main gear.
The gear will fall under gravity (more accurately aircraft vertical G)
When all gear is locked the three green lights will come on.
Attempting to land before the gear is locked will result in the gear failing (retracting).
It may be necessary to give the aircraft a little + or - g to get it locked.

CADC warnings for landing gear (the tone) operate the same, so if not locked within ten seconds of deployment a "beep-beep-beep" tone (same as the landing gear not down tone) will be heard until the gear is locked.

If electrics and hydraulics are present it is possible to retract the gear after emergency deployment, otherwise the gear will stay down.