F-15: Radar optimisations.
Changed the following. This is roughly in order of the most to the least importance in terms of run time efficiency.

1 - only scan the ai/models when it is changed. This builds tgt_list. This could possibly be further optimised to avoid GC; but actually it's not called that often anymore.
2 - reorder the evaluation of the is visible logic so that it does the least costly first (range, az) and then the RCS and terrain. It does RCS first because that looks to me to be the lighter of the two.
3 - do not call the get_cart_ground_intersection (or the other method) when both alt above 8900
4 - divide the update scan across multiple loops. Because tgt_list is in distance order we can simply set the rest of the targets to be not displayed and break out of the loop.
5 - took out the ARA-63 stuff.
6 - Always keep tgt_list sorted by distance so we can break out of the target update the moment that the range is exceeded. This means it doesn't have to be sorted for next/prev.
7 - remove getprop and replace with props.getNode

Prior to these optimisations the radar could consume 42fps (drop from 55 to 12). see https://i.imgur.com/zw18Gre.png

Also this removes the outdated ARA-63 that used the AI tree. This is replaced with the Emesary version. Not very relevant for this aircraft as it isn't carrier capable. Something will have to be added back into the F-14 so that the ground services dialog can allow carrier repositions - but that can be a much simpler implementation.