F-14 simplify VDI model
The VDI ground section was inclined - I assume to provide a more realistic view. However this is unnecessary as the texture translation provides a sufficiently accurate instrument and the inclined view is only valid given a limited range of pilot head movement.

The ground section is now a polygon in the same plane as the actual VDI CRT display.

Also add section to the base cockpit to fill the gap between the VDI and HSI that previously allowed parts of the VDI display that should be clipped (but weren't) to be visible.