F-14: External stores model unification.
Move all of the external stores into a single 3d model, with a single texture atlas.

The submodel for external tanks also uses the same texture.

This is partly for efficiency but also part of the process of moving the texture for the external stores into the livery.

(cherry picked from commit 063388e9f95c3ff4962edd6ac1a527dd5aa949f7)