F-15: PTC / Pitch CAS changes
Apply trim to raw stick position as this will allow trim to command +/-G values - effectively trim works on the stick position which is in accordance with the docs.

Reduce CAS PTC authority at lower speeds; this is something that I've possibly invented - but without this CAS tends to pitch up quite a lot at low speeds and this leads to strange handling with often forward stick pressure required to land. Reducing the authority makes it fly *subjectively* better.