Tuning of flaps
This is a revisit of the alpha / approach handling - based on the discussion with timi[1].

I've revisited all of the relevant literature and confirmed the following;

1. Indicated alpha is probably about right
2. Approach speeds too high because available lift is possibly a little low
3. Too much pitch because of (2)

For comparison purposes the Heatblur F-14B performance was measured during case 1 recovery and revealed effects consistent with more lift from the flaps, slightly better pitch stability at high alpha and consequently easier to hold onspeed. Required elevator was measured and was lower and for a similar configuration. My view is that confirming that (2) is probably the root cause; i.e. not enough lift with full flaps and that the best approach is to tune the lift and then balance out the rest of the coefficients.

HB approach figures as measured[2]
* Pitch angle   ~8.5 degrees
* Elevator      ~9.5 degrees elevator
* Alpha         ~15 units indicated
* Speed         ~124kts
* Fuel          ~6000lbs
* Config         External tanks only

The changes I've made based on the investigation both static calculations and using OpenVSP derived data gives the following results
1. C.G moved a lot closer AERORP; this is consistent with aerodynamic theory
2. Airspeed / lift showed that a factor of ~1.5 (different values for main and auxiliary flaps) gives better speeds.
3. Drag adjusted to give reasonable power settings; factor of ~1.3
4. Pitching moment too much nose down, conclusion is that the main flaps must reverse when full flaps are deployed otherwise far too much elevator is required (~15 degrees)

[1] http://fguk.me/forum/development-hangar/8371-f-14-aoa-indexer#45576
[2] https://i.imgur.com/aXvizJU.png