SAS roll / yaw fixes
Reintroduce roll-sas-2 after correcting errors related to sign conventions. This is the roll augmentation either fixed (for high alpha) or a lead/lag version for lower alphas (with a blend of the two beetween 10 and 21 alpha units). Also added use a force to position adjustment for stick roll position as the controls on the aircraft have much heavier loading than a joystick spring.

Fixed SAS Yaw 2 (stick) as it wasn't using the right eqn or scaling. Fixed sas-yaw-p as not using correct scaling and also using r instead of p.