F-14: Fix red flood and instruments
The red and white flood and instrument lighting wheel now work together so that the overall illumination will be blended from the lights.

When the red flood is on dim there will be dim light on mostly everything; unless the instrument wheel is set to a value that overpowers the red light; equally the white flood light will either be dim, bright or off and this will generally make most things white.