F-14 aero; revised flaps/slats
New tables for main flaps, aux flaps, slats for lift, drag and pitch (longitudinal) - this is based on the only figures that I have reliable data for which is the increment due to full flaps and slats.

The approach taken is to allocate from the total increment 20% to slats (17 degrees) and 80% to flaps (at 35 degrees). The 80% of the flaps total is then split out so that 30% goes to the AUX flaps and 70% to the main flaps.

This ratios are based on other aircraft data and until I find some better data we may have to accept that for intermediate values of flaps and slats there may be around a 10% error.

Overall error is within 2% of published; Figure 11-8. Landing Approach Airspeed (15 Units AOA) in NAVAIR 01-F14AAP-1, also in NAVAIR 01-F14AAP-1.1 Figure 7-1 Approach Speed DLC netural- Flaps down (15 units AOA)  two engine gives us 136 kts IAS; which when allowing 8knots for DLC gives us also 124 kts IAS.