F-14: SAS changes for roll and yaw
Yaw channel was using incorrect units for the feedback from
SAS into the main rudder command; with a large over control
as a result. Changed the rudder to work in degrees rather than
normalized.

The net result of the use of the incorrect units
(and magnitude) was to in effect render the yaw SAS ineffective
and this change now fixes it.

Move rudder control into yaw channel (previously was in its
own JSBSim channel)

Roll channel removed the stick actuator (rate limit) and the '
Stick position force simulation (based on NAVAIR 01-F14AAP-1,
2.22.3.1 Lateral Feel)