F-14 : F110-GE-400 fuel and oil pressure and temperature
Added a basic simulation of the fuel temperature and pressure and
the oil temperature and pressure.

Fuel temperature is basically the same as ambient with a small
delta to account for heating due to fuel pumps.

Fuel pressure is a function of N2.

Oil temperature is built up from 3 elements with the oil temperature
being related to fuel flow because the fuel is used to cool the oil.

This gives us the following.

* Base temperature (from N2 and fuel flow)
* N2 increment
* high speed increment

This simulation is based on my notes taken from a simulator session
and reworked into the above.  On reflection the entire three stage
calculation may be flawed because we're using N2 twice - so it will possibly
get modified shortly.