F-14: rework oil temperature and pressure
Based on the same data but reworked to be more sensibly presented.

The basic oil temperature is the best I can find; it's effectively
a linear function of fuel flow and altitude. Possibly this isn't correct
but it is the only data I have.

The oil pressure is based on oil temperature and N2%, normalised and then
denormalized by a fixed factor that is be slightly different for each engine.
The data for this table is extrapolated from the data that I did have but it does
produce the right results at the datapoints.

Overall this is an improvement on the previous model - but maybe further refinement
will be needed