F-14 ; new simulation of the alpha measurement devices,
The aircraft has 3 alpha measuring devices of these; one on each side near the cockpit and one on the nose boom.

As of 2021-07-14 I have not managed to find any indication of how these are handled so in abscence of real data I have devised a plausible system based in part on data from the F-18

The way I've implemented it is basically to limit the input, added a lead leag compensator (from the F-18 report) and to fade it to zero based on qbar so that on ground it will effectively be reading zero (because there isn't enough air to be detected).

* alpha-indicated-deg-0 - alpha units (y=mx+c)
* alpha-indicated-deg-1 - lead lag compensator
                        _______1_______
                            0.073s + 1

* alpha-indicated-deg  - adjusted to take airflow into account