mp-network.nas: Declare 'str' as variable
In FG 2020.4, str() in the global nasal namespace is a function to
convert to strings. Assigning to 'str' without declaring it as variable

    str = [...]

overwrites this global function, affecting all namespaces.
Since mp-network.nas is loaded from the model xml file, this can also
happen when loading the F-15 as a MP model.  This caused errors in the
Viggen which uses the str() function.

Change the assignment to keep the variable local:

    var str = [...]