F-15: fix AAR bug
Need to mark the feed tanks as not selected as otherwise the core AAR seems to have a bug where it will continually reset the level on these tanks which somehow overrides the consumption.

Quite why fgdata AAR is managing fuel levels like this is a bit of a mystery to me as it doesn't seem right but probably is OK for a generic system.